# docs/openapi/zerg.v1.json
{
  "components": {
...
    "schemas": {
...
      "CreateOrgBody": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "InviteBody": {
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
...
      "MemberResponse": {
        "description": "An org member (WorkOS-shaped; local-only for personal workspaces).",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "workos_user_id": {
            "type": "string"
          }
        },
        "required": [
          "workos_user_id",
          "email",
          "name",
          "role",
          "status"
        ],
        "type": "object"
      },
...
      "OrgResponse": {
        "description": "The caller's active organization.",
        "properties": {
          "external_id": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "is_personal": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "external_id",
          "name",
          "role",
          "is_personal"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "in": "cookie",
        "name": "zerg_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API for managing tasks, projects, cloud resources, and users",
    "license": {
      "name": ""
    },
    "title": "Zerg API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/callback": {
      "get": {
        "operationId": "callback",
        "parameters": [
          {
            "description": "Authorization code from WorkOS",
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "CSRF state echoed by WorkOS",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Error code if WorkOS rejected the request",
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Session established; redirect to the frontend (sets session + CSRF cookies)"
          },
          "400": {
            "description": "Missing/expired flow, missing code/state, or state mismatch"
          }
        },
        "summary": "`GET /api/auth/callback` → exchange code, provision the user, mint a server-side session.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/org": {
      "get": {
        "operationId": "get_org",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgResponse"
                }
              }
            },
            "description": "Active organization"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "`GET /api/org` → the caller's active org context.",
        "tags": [
          "org"
        ]
      },
      "post": {
        "operationId": "create_org",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgResponse"
                }
              }
            },
            "description": "Organization created; session switched"
          },
          "400": {
            "description": "Empty name, or caller already belongs to an organization"
          },
          "502": {
            "description": "WorkOS error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "`POST /api/org` → self-serve org creation: create the WorkOS org + admin\nmembership, then re-mint the current session for the new org (org-switching\nrefresh grant) so no re-login is needed.",
        "tags": [
          "org"
        ]
      }
    },
    "/org/invitations": {
      "get": {
        "operationId": "list_invitations",
        "responses": {
          "200": {
            "description": "Org invitations"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "`GET /api/org/invitations` → pending/recent invitations (empty for personal).",
        "tags": [
          "org"
        ]
      },
      "post": {
        "operationId": "create_invitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Invitation created"
          },
          "400": {
            "description": "Personal workspaces cannot invite"
          },
          "403": {
            "description": "Only org admins can invite"
          },
          "502": {
            "description": "WorkOS error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "`POST /api/org/invitations` → invite `email` to the org (admin only). WorkOS\nsends the email and hosts the accept flow.",
        "tags": [
          "org"
        ]
      }
    },
    "/org/invitations/{id}/revoke": {
      "post": {
        "operationId": "revoke_invitation",
        "parameters": [
          {
            "description": "WorkOS invitation id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation revoked"
          },
          "403": {
            "description": "Only org admins can revoke"
          },
          "502": {
            "description": "WorkOS error"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "`POST /api/org/invitations/{id}/revoke` → revoke a pending invitation (admin only).",
        "tags": [
          "org"
        ]
      }
    },
    "/org/members": {
      "get": {
        "operationId": "list_members",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Org members"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "`GET /api/org/members` → the org's members. Personal workspaces contain only\nthe owner; real orgs join WorkOS memberships (roles) with WorkOS users (emails).",
        "tags": [
          "org"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "API base path",
      "url": "/api"
    }
  ],
...
}